# api/openapi/maia.yaml
openapi: 3.1.0
info:
  title: MAIA API
  description: |
    MAIA (Memory AI Architecture) is an AI-native distributed memory system for LLM context management.

    ## Authentication

    MAIA supports API key authentication. Include your API key in requests using one of these methods:
    - Header: `X-API-Key: your-api-key`
    - Header: `Authorization: Bearer your-api-key`
    - Query parameter: `?api_key=your-api-key`

    ## Scopes

    API keys can be restricted to specific operations using scopes. Available scopes:

    | Scope | Description |
    |-------|-------------|
    | `*` | Full access (wildcard) |
    | `read` | Read memories and namespaces |
    | `write` | Create and update memories and namespaces |
    | `delete` | Delete memories and namespaces |
    | `search` | Search operations |
    | `context` | Context assembly operations |
    | `inference` | Inference operations |
    | `stats` | View statistics and metrics |
    | `admin` | Administrative operations (tenant/API key management) |

    API keys with no scopes (empty) are granted full access for backward compatibility.

    When a request lacks the required scope, a `403 Forbidden` response is returned with code `INSUFFICIENT_SCOPE`.

    ## Rate Limiting

    API requests are rate limited. When you exceed the limit, you'll receive a `429 Too Many Requests` response.

    ## Multi-Tenancy

    MAIA supports multi-tenancy with tenant isolation. Use the `X-MAIA-Tenant-ID` header to specify
...

...
  license:
...
servers:
...
paths:
...
  /v1/memories:
    post:
      summary: Create a memory
      description: Creates a new memory in the specified namespace
      operationId: createMemory
      tags:
        - memories
      x-required-scopes:
        - write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMemoryRequest'
      responses:
        '201':
          description: Memory created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
  /v1/memories/{id}:
    get:
      summary: Get a memory
      description: Retrieves a memory by its ID
      operationId: getMemory
      tags:
        - memories
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '200':
          description: Memory retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a memory
      description: Updates an existing memory
      operationId: updateMemory
      tags:
        - memories
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemoryRequest'
      responses:
        '200':
          description: Memory updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Memory'
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a memory
      description: Deletes a memory by its ID
      operationId: deleteMemory
      tags:
        - memories
      parameters:
        - $ref: '#/components/parameters/MemoryId'
      responses:
        '204':
          description: Memory deleted successfully
        '404':
          description: Memory not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/memories/search:
    post:
      summary: Search memories
      description: Searches for memories matching the query
      operationId: searchMemories
      tags:
        - memories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchMemoriesRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchMemoriesResponse'
  /v1/namespaces:
    get:
      summary: List namespaces
      description: Returns a list of all namespaces
      operationId: listNamespaces
      tags:
        - namespaces
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of namespaces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListNamespacesResponse'

    post:
      summary: Create a namespace
      description: Creates a new namespace
      operationId: createNamespace
      tags:
        - namespaces
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNamespaceRequest'
      responses:
        '201':
          description: Namespace created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Namespace already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/namespaces/{id}:
    get:
      summary: Get a namespace
      description: Retrieves a namespace by its ID
      operationId: getNamespace
      tags:
        - namespaces
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
      responses:
        '200':
          description: Namespace retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a namespace
      description: Updates an existing namespace
      operationId: updateNamespace
      tags:
        - namespaces
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNamespaceRequest'
      responses:
        '200':
          description: Namespace updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Namespace'
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a namespace
      description: Deletes a namespace and all its memories
      operationId: deleteNamespace
      tags:
        - namespaces
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
      responses:
        '204':
          description: Namespace deleted successfully
        '404':
          description: Namespace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/namespaces/{id}/memories:
    get:
      summary: List namespace memories
      description: Returns all memories in a namespace
      operationId: listNamespaceMemories
      tags:
        - namespaces
        - memories
      parameters:
        - $ref: '#/components/parameters/NamespaceId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of memories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListMemoriesResponse'
  /v1/context:
    post:
      summary: Assemble context
      description: |
        Assembles context from memories for a given query.
        Uses position-aware assembly to optimize context for LLM consumption.
      operationId: getContext
      tags:
        - context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetContextRequest'
      responses:
        '200':
          description: Assembled context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextResponse'
...
  /admin/tenants:
    post:
      summary: Create a tenant
      description: Creates a new tenant (requires admin scope)
      operationId: createTenant
      tags:
        - admin
      x-required-scopes:
        - admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
        '409':
          description: Tenant already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List tenants
      description: Returns a list of all tenants (requires admin scope)
      operationId: listTenants
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TenantStatus'
        - name: plan
          in: query
          schema:
            $ref: '#/components/schemas/TenantPlan'
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTenantsResponse'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
  /admin/tenants/{id}:
    get:
      summary: Get a tenant
      description: Retrieves a tenant by ID (requires admin scope)
      operationId: getTenant
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update a tenant
      description: Updates an existing tenant (requires admin scope)
      operationId: updateTenant
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenantRequest'
      responses:
        '200':
          description: Tenant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Delete a tenant
      description: Deletes a tenant (requires admin scope)
      operationId: deleteTenant
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '204':
          description: Tenant deleted successfully
        '403':
          description: Insufficient scope or system tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/tenants/{id}/usage:
    get:
      summary: Get tenant usage
      description: Returns usage statistics for a tenant (requires admin scope)
      operationId: getTenantUsage
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant usage statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantUsage'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/tenants/{id}/suspend:
    post:
      summary: Suspend a tenant
      description: Suspends a tenant (requires admin scope)
      operationId: suspendTenant
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant suspended successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          description: Insufficient scope or system tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/tenants/{id}/activate:
    post:
      summary: Activate a tenant
      description: Activates a suspended tenant (requires admin scope)
      operationId: activateTenant
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: Tenant activated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
        '404':
          description: Tenant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/apikeys/{key}:
    delete:
      summary: Revoke an API key
      description: Revokes an API key (requires admin scope)
      operationId: revokeAPIKey
      tags:
        - admin
      x-required-scopes:
        - admin
      parameters:
        - name: key
          in: path
          required: true
          description: API key hash
          schema:
            type: string
      responses:
        '204':
          description: API key revoked successfully
        '403':
          description: Insufficient scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScopeError'
        '404':
          description: API key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  parameters:
    MemoryId:
      name: id
      in: path
      required: true
      description: Memory ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateMemoryRequest:
      type: object
      properties:
        namespace:
          type: string
          description: Target namespace
        content:
          type: string
          description: Memory content
          minLength: 1
          maxLength: 10000
        type:
          $ref: '#/components/schemas/MemoryType'
        metadata:
          type: object
          additionalProperties: true
        tags:
          type: array
          items:
            type: string
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          default: 1.0
        source:
          $ref: '#/components/schemas/MemorySource'
      required:
        - namespace
        - content
...
    SearchMemoriesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScoredMemory'
        count:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: string
          description: Additional details
      required:
        - error
...
security:
  - ApiKeyHeader: []
  - BearerAuth: []